Separate internal logic for search actions into files
