+++ src/actions/search/procs.js
@@ -1,0 +1,6 @@
+import { searchProcFactory } from './utils';
+
+export const ActivitySearchProc = searchProcFactory('activity');
+export const PersonSearchProc = searchProcFactory('person');
+export const PersonQuerySearchProc = searchProcFactory('personquery');
+export const SurveySubmissionSearchProc = searchProcFactory('surveysubmission');
